Fix broken unit test
An nunit test [failed on the build server](http://mp-ci.centralus.cloudapp.azure.com/viewLog.html?buildId=34105&buildTypeId=MicroClients_Integration_BuildAndTest_CrdsAngular&tab=buildLog#_focus=17699), but worked successfully on the next run, without changes.  Further investigation showed that the way the test was written, a DateTime expectation could occasionally fail if the test fixture did not execute within the same second the expectation was set.  Updated the test (and others with the same issue) to consider the test successful if the expected date is within 60 seconds of the actual date.