Setting Program end dates to 2 years out
Test pledges were not avaialable to give to. This is because thier
end date was last year. The SQL script to create the pledges and
programs was setting all programs to end at the end of the year.
This SQL script is not ran very often and usually never ran at the
begining of the year. So I changed the script to always end date the
programs two years 2 years out.